Remodified Tree's logic

+++ Game/Tree.h
@@ -2,7 +2,7 @@
 //  Tree.h
 //  
 //  Represents the Tree of States for AI Algorithms
-//
+//  Represented as a sort of linked list with vector traversal??
 
 #ifndef State_h
 #define State_h
@@ -27,10 +27,9 @@ public:
     ~Tree(){}
     	
 	Node get_tree(); //Not sure on the data type for this
-	void add_node_w_children(vector<State> children, vector<int> value, Node* parent_node/*Or just Node??*/); //Add new_node to node_list
+	void add_node_w_children(vector<State> children, vector<long int> value, Node* parent_node/*Or just Node??*/); //Add new_node to node_list
 	//void add_node_w_children(Node new_node);//Not sure which one should be implemented??
 	Node get_node(int depth/*SOMETHING ELSE IS NEEDED TOO*/);
-	
 	
 };
